refactor registries
