+++ common/src/main/java/com/dfsek/terra/api/structures/script/StructureScript.java
@@ -28,8 +28,8 @@ import com.dfsek.terra.api.structures.structure.Rotation;
 import com.dfsek.terra.api.structures.structure.buffer.Buffer;
 import com.dfsek.terra.api.structures.structure.buffer.DirectBuffer;
 import com.dfsek.terra.api.structures.structure.buffer.StructureBuffer;
-import com.dfsek.terra.registry.LootRegistry;
-import com.dfsek.terra.registry.ScriptRegistry;
+import com.dfsek.terra.registry.config.LootRegistry;
+import com.dfsek.terra.registry.config.ScriptRegistry;
 import com.dfsek.terra.world.generation.math.SamplerCache;
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
