+++ common/src/main/java/com/dfsek/terra/config/loaders/config/biome/templates/source/ImageProviderTemplate.java
@@ -3,7 +3,7 @@ package com.dfsek.terra.config.loaders.config.biome.templates.source;
 import com.dfsek.tectonic.annotations.Value;
 import com.dfsek.terra.biome.provider.BiomeProvider;
 import com.dfsek.terra.biome.provider.ImageBiomeProvider;
-import com.dfsek.terra.registry.BiomeRegistry;
+import com.dfsek.terra.registry.config.BiomeRegistry;
 
 import java.awt.image.BufferedImage;
