+++ common/src/main/java/com/dfsek/terra/config/pack/ConfigPack.java
@@ -42,16 +42,16 @@ import com.dfsek.terra.config.templates.OreTemplate;
 import com.dfsek.terra.config.templates.PaletteTemplate;
 import com.dfsek.terra.config.templates.StructureTemplate;
 import com.dfsek.terra.config.templates.TreeTemplate;
-import com.dfsek.terra.registry.BiomeRegistry;
-import com.dfsek.terra.registry.CarverRegistry;
-import com.dfsek.terra.registry.FloraRegistry;
-import com.dfsek.terra.registry.LootRegistry;
-import com.dfsek.terra.registry.OreRegistry;
-import com.dfsek.terra.registry.PaletteRegistry;
-import com.dfsek.terra.registry.ScriptRegistry;
-import com.dfsek.terra.registry.StructureRegistry;
 import com.dfsek.terra.registry.TerraRegistry;
-import com.dfsek.terra.registry.TreeRegistry;
+import com.dfsek.terra.registry.config.BiomeRegistry;
+import com.dfsek.terra.registry.config.CarverRegistry;
+import com.dfsek.terra.registry.config.FloraRegistry;
+import com.dfsek.terra.registry.config.LootRegistry;
+import com.dfsek.terra.registry.config.OreRegistry;
+import com.dfsek.terra.registry.config.PaletteRegistry;
+import com.dfsek.terra.registry.config.ScriptRegistry;
+import com.dfsek.terra.registry.config.StructureRegistry;
+import com.dfsek.terra.registry.config.TreeRegistry;
 import com.dfsek.terra.world.generation.math.SamplerCache;
 import com.dfsek.terra.world.population.items.TerraStructure;
 import com.dfsek.terra.world.population.items.ores.Ore;
