+++ common/src/main/java/com/dfsek/terra/registry/config/CarverRegistry.java
@@ -1,6 +1,7 @@
-package com.dfsek.terra.registry;
+package com.dfsek.terra.registry.config;
 
 import com.dfsek.terra.carving.UserDefinedCarver;
+import com.dfsek.terra.registry.TerraRegistry;
 
 public class CarverRegistry extends TerraRegistry<UserDefinedCarver> {
 }
