+++ common/src/main/java/com/dfsek/terra/registry/config/FloraRegistry.java
@@ -1,9 +1,10 @@
-package com.dfsek.terra.registry;
+package com.dfsek.terra.registry.config;
 
 import com.dfsek.terra.api.core.TerraPlugin;
 import com.dfsek.terra.api.platform.block.BlockData;
 import com.dfsek.terra.api.platform.block.MaterialData;
 import com.dfsek.terra.api.world.flora.Flora;
+import com.dfsek.terra.registry.TerraRegistry;
 import com.dfsek.terra.util.MaterialSet;
 import com.dfsek.terra.world.population.items.flora.ConstantFlora;
