+++ common/src/main/java/com/dfsek/terra/registry/config/LootRegistry.java
@@ -1,6 +1,7 @@
-package com.dfsek.terra.registry;
+package com.dfsek.terra.registry.config;
 
 import com.dfsek.terra.api.structures.loot.LootTable;
+import com.dfsek.terra.registry.TerraRegistry;
 
 public class LootRegistry extends TerraRegistry<LootTable> {
 }
