+++ common/src/main/java/com/dfsek/terra/registry/config/ScriptRegistry.java
@@ -1,6 +1,7 @@
-package com.dfsek.terra.registry;
+package com.dfsek.terra.registry.config;
 
 import com.dfsek.terra.api.structures.script.StructureScript;
+import com.dfsek.terra.registry.TerraRegistry;
 
 public class ScriptRegistry extends TerraRegistry<StructureScript> {
 }
