+++ common/src/main/java/com/dfsek/terra/registry/config/StructureRegistry.java
@@ -1,5 +1,6 @@
-package com.dfsek.terra.registry;
+package com.dfsek.terra.registry.config;
 
+import com.dfsek.terra.registry.TerraRegistry;
 import com.dfsek.terra.world.population.items.TerraStructure;
 
 public class StructureRegistry extends TerraRegistry<TerraStructure> {
