+++ common/src/test/java/biome/DistributionTest.java
@@ -33,8 +33,8 @@ import com.dfsek.terra.config.loaders.config.sampler.NoiseSamplerBuilderLoader;
 import com.dfsek.terra.config.pack.ConfigPack;
 import com.dfsek.terra.config.templates.AbstractableTemplate;
 import com.dfsek.terra.debug.DebugLogger;
-import com.dfsek.terra.registry.BiomeRegistry;
 import com.dfsek.terra.registry.ConfigRegistry;
+import com.dfsek.terra.registry.config.BiomeRegistry;
 import com.dfsek.terra.world.TerraWorld;
 
 import javax.swing.*;
